+++ modules/runtime/include/utils/parallel_for.hpp
@@ -11,7 +11,7 @@ void parallel_for(Iter begin, Iter end, size_t batch, F f)
     typename std::iterator_traits<Iter>::difference_type n = std::distance(begin, end);
     size_t batchCount = (n / batch) + 1;
     task::counter_t counter;
-    counter.add(batchCount);
+    counter.add((uint32_t)batchCount);
     for (size_t i = 0; i < batchCount; ++i)
     {
         auto toAdvance = std::min((size_t)n, batch);
